requirements.txt
beautifulsoup4==4.11.1
cloudscraper==1.2.60
discord.py==1.7.3
